AAE5303_UNet_demo_code_only/requirements.txt
# AAE5303 - Semantic Segmentation Assignment
# Python dependencies for training, evaluation and visualization

# Core dependencies
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pillow>=9.0.0

# Deep learning framework
torch>=2.0.0
torchvision>=0.15.0

# Training utilities
tqdm>=4.64.0
wandb>=0.12.0

# For metric calculation (leaderboard submission)
scikit-image>=0.19.0
opencv-python>=4.5.0
